Compile solver to wasm
This commit introduces a new crate, solver-wasm, that uses wasm-bindgen
to expose the solver::solve method to javascript via WebAssembly.

The SolveResult that we pass from rust to js is serialized to JSON via
Serde. This made for easier handling of multiple return values (both an
optional error and an optional solution) while still using a JsValue. A
JsValue, as opposed to returning a rust struct directly, is memory
managed automatically (as far as I've understood).